Refactored as per Clippy. Modified Population::evaluate_fitness to take any type E: Fn. Population::evolve no longer automatically resets on population degeneration. Instead, an error is returned. Population::reset is now public. Elitism is now a flat number, instead of a percentage.

+++ neat/src/populations/config.rs
@@ -8,9 +8,9 @@ pub struct PopulationConfig {
     /// genomes are considered as belonging to
     /// different species.
     pub distance_threshold: f32,
-    /// Top % of each species which is copied
+    /// Top n of each species which is copied
     /// as-is to the next generation.
-    pub elitism: f32,
+    pub elitism: usize,
     /// Top % of each species which can participate
     /// in mating.
     pub survival_threshold: f32,
